Moved services.ts and tests to metadata-common
