+++ packages/application/src/index.ts
@@ -17,5 +17,4 @@
 import '../style/index.css';
 
 export * from './parsing';
-export * from './services';
 export * from './requests';
