+++ packages/application/src/test/application.spec.ts
@@ -20,7 +20,6 @@ import { JupyterServer, NBTestUtils } from '@jupyterlab/testutils';
 
 import { NotebookParser } from '../parsing';
 import { RequestHandler } from '../requests';
-import { FrontendServices } from '../services';
 
 const server = new JupyterServer();
 const notebookWithEnvVars: any = {
@@ -83,79 +82,6 @@ afterAll(async () => {
 });
 
 describe('@elyra/application', () => {
-  describe('FrontendServices', () => {
-    describe('#getSchema', () => {
-      it('should get schema', async () => {
-        const schemaResponse = await FrontendServices.getSchema(
-          'code-snippets'
-        );
-        expect(schemaResponse[0]).toHaveProperty(
-          'properties.schema_name.description',
-          'The schema associated with this instance'
-        );
-      });
-    });
-    describe('#getAllSchema', () => {
-      it('should get all schema', async () => {
-        const schemas = await FrontendServices.getAllSchema();
-        const schemaNames = schemas.map((schema: any) => {
-          return schema.name;
-        });
-        const knownSchemaNames = ['code-snippet', 'runtime-image', 'kfp'];
-        for (const schemaName of knownSchemaNames) {
-          expect(schemaNames).toContain(schemaName);
-        }
-        expect(schemas.length).toBeGreaterThanOrEqual(knownSchemaNames.length);
-      });
-    });
-    describe('metadata requests', () => {
-      beforeAll(async () => {
-        const existingSnippets = await FrontendServices.getMetadata(
-          'code-snippets'
-        );
-        if (
-          existingSnippets.find((snippet: any) => {
-            return snippet.name === 'tester';
-          })
-        ) {
-          await FrontendServices.deleteMetadata('code-snippet', 'tester');
-        }
-      });
-
-      it('should create metadata instance', async () => {
-        expect(
-          await FrontendServices.postMetadata(
-            'code-snippets',
-            JSON.stringify(codeSnippetMetadata)
-          )
-        ).toMatchObject(codeSnippetMetadata);
-      });
-
-      it('should get the correct metadata instance', async () => {
-        expect(
-          await FrontendServices.getMetadata('code-snippets')
-        ).toContainEqual(codeSnippetMetadata);
-      });
-
-      it('should update the metadata instance', async () => {
-        codeSnippetMetadata.metadata.code = ['testing'];
-        expect(
-          await FrontendServices.putMetadata(
-            'code-snippets',
-            'tester',
-            JSON.stringify(codeSnippetMetadata)
-          )
-        ).toHaveProperty('metadata.code', ['testing']);
-      });
-
-      it('should delete the metadata instance', async () => {
-        await FrontendServices.deleteMetadata('code-snippets', 'tester');
-        const snippets = await FrontendServices.getMetadata('code-snippets');
-        expect(snippets).not.toContain(codeSnippetMetadata);
-      });
-    });
-  });
-
   describe('RequestHandler', () => {
     describe('#makeGetRequest', () => {
       it('should make get request', async () => {
