+++ packages/metadata-common/package.json
@@ -27,6 +27,7 @@
     "clean": "rimraf lib",
     "dist": "npm pack .",
     "prepare": "npm run build",
+    "test": "jest",
     "watch": "tsc -w"
   },
   "dependencies": {
@@ -43,7 +44,21 @@
     "react": "^16.8.6"
   },
   "devDependencies": {
+    "@types/enzyme": "^3.10.5",
+    "@types/enzyme-adapter-react-16": "^1.0.6",
+    "@types/jest": "^23.3.11",
+    "@types/react": "^16.8.23",
+    "@types/react-dom": "^16.8.5",
+    "@types/react-intl": "^2.3.0",
+    "enzyme": "^3.11.0",
+    "enzyme-adapter-react-16": "^1.15.3",
+    "install": "^0.13.0",
+    "jest": "^24.7.1",
+    "jest-raw-loader": "^1.0.1",
     "rimraf": "~3.0.0",
+    "source-map-loader": "^0.2.4",
+    "ts-jest": "^24.0.2",
+    "ts-loader": "^6.2.1",
     "typescript": "~3.7.3"
   },
   "publishConfig": {
