+++ packages/metadata-common/tsconfig.test.json
@@ -1,0 +1,5 @@
+{
+  "extends": "../../tsconfigbase.test",
+  "include": ["src/*", "test/*"],
+  "references": []
+}
